refactor: extract SubsonicSearchController — fixes #85
