+++ airsonic-main/src/main/java/org/airsonic/player/controller/SubsonicRESTController.java
@@ -30,7 +30,6 @@ import org.airsonic.player.domain.User;
 import org.airsonic.player.i18n.LocaleResolver;
 import org.airsonic.player.service.*;
 import org.airsonic.player.service.podcast.PodcastDownloadClient;
-import org.airsonic.player.service.search.IndexType;
 import org.airsonic.player.util.NetworkUtil;
 import org.airsonic.player.util.StringUtil;
 import org.airsonic.player.util.Util;
@@ -125,8 +124,6 @@ public class SubsonicRESTController {
     private PodcastDownloadClient podcastDownloadClient;
     @Autowired
     private RatingService ratingService;
-    @Autowired
-    private SearchService searchService;
     @Autowired
     private ArtistService artistService;
     @Autowired
@@ -338,148 +335,6 @@ public class SubsonicRESTController {
 
 
 
-
-    @RequestMapping({"/search", "/search.view"})
-    public void search(HttpServletRequest request, HttpServletResponse response) throws Exception {
-        request = wrapRequest(request);
-        String username = securityService.getCurrentUsername(request);
-        Player player = playerService.getPlayer(request, response, username);
-
-        String any = request.getParameter("any");
-        String artist = request.getParameter("artist");
-        String album = request.getParameter("album");
-        String title = request.getParameter("title");
-
-        StringBuilder query = new StringBuilder();
-        if (any != null) {
-            query.append(any).append(" ");
-        }
-        if (artist != null) {
-            query.append(artist).append(" ");
-        }
-        if (album != null) {
-            query.append(album).append(" ");
-        }
-        if (title != null) {
-            query.append(title);
-        }
-
-        SearchCriteria criteria = new SearchCriteria();
-        criteria.setQuery(query.toString().trim());
-        criteria.setCount(getIntParameter(request, "count", 20));
-        criteria.setOffset(getIntParameter(request, "offset", 0));
-        List<org.airsonic.player.domain.MusicFolder> musicFolders = mediaFolderService.getMusicFoldersForUser(username);
-
-        org.airsonic.player.domain.SearchResult result = searchService.search(criteria, musicFolders, IndexType.SONG);
-        org.subsonic.restapi.SearchResult searchResult = new org.subsonic.restapi.SearchResult();
-        searchResult.setOffset(result.getOffset());
-        searchResult.setTotalHits(result.getTotalHits());
-
-        for (MediaFile mediaFile : result.getMediaFiles()) {
-            searchResult.getMatch().add(jaxbContentService.createJaxbChild(player, mediaFile, username));
-        }
-        Response res = createResponse();
-        res.setSearchResult(searchResult);
-        jaxbWriter.writeResponse(request, response, res);
-    }
-
-    @RequestMapping({"/search2", "/search2.view"})
-    public void search2(HttpServletRequest request, HttpServletResponse response) throws Exception {
-        request = wrapRequest(request);
-        String username = securityService.getCurrentUsername(request);
-        Player player = playerService.getPlayer(request, response, username);
-        Integer musicFolderId = getIntParameter(request, "musicFolderId");
-        List<org.airsonic.player.domain.MusicFolder> musicFolders = mediaFolderService.getMusicFoldersForUser(username, musicFolderId);
-
-        SearchResult2 searchResult = new SearchResult2();
-
-        String query = request.getParameter("query");
-        SearchCriteria criteria = new SearchCriteria();
-        criteria.setQuery(StringUtils.trimToEmpty(query));
-        criteria.setCount(getIntParameter(request, "artistCount", 20));
-        criteria.setOffset(getIntParameter(request, "artistOffset", 0));
-        org.airsonic.player.domain.SearchResult artists = searchService.search(criteria, musicFolders, IndexType.ARTIST);
-        for (MediaFile mediaFile : artists.getMediaFiles()) {
-            searchResult.getArtist().add(jaxbContentService.createJaxbArtist(mediaFile, username));
-        }
-
-        criteria.setCount(getIntParameter(request, "albumCount", 20));
-        criteria.setOffset(getIntParameter(request, "albumOffset", 0));
-        org.airsonic.player.domain.SearchResult albums = searchService.search(criteria, musicFolders, IndexType.ALBUM);
-        for (MediaFile mediaFile : albums.getMediaFiles()) {
-            searchResult.getAlbum().add(jaxbContentService.createJaxbChild(player, mediaFile, username));
-        }
-
-        criteria.setCount(getIntParameter(request, "songCount", 20));
-        criteria.setOffset(getIntParameter(request, "songOffset", 0));
-        org.airsonic.player.domain.SearchResult songs = searchService.search(criteria, musicFolders, IndexType.SONG);
-        for (MediaFile mediaFile : songs.getMediaFiles()) {
-            searchResult.getSong().add(jaxbContentService.createJaxbChild(player, mediaFile, username));
-        }
-
-        Response res = createResponse();
-        res.setSearchResult2(searchResult);
-        jaxbWriter.writeResponse(request, response, res);
-    }
-
-    @RequestMapping({"/search3", "/search3.view"})
-    public void search3(HttpServletRequest request, HttpServletResponse response) throws Exception {
-        request = wrapRequest(request);
-        String username = securityService.getCurrentUsername(request);
-        Player player = playerService.getPlayer(request, response, username);
-        Integer musicFolderId = getIntParameter(request, "musicFolderId");
-        List<org.airsonic.player.domain.MusicFolder> musicFolders = mediaFolderService.getMusicFoldersForUser(username, musicFolderId);
-
-        SearchResult3 searchResult = new SearchResult3();
-
-        String query = request.getParameter("query");
-        // replace empty string with null
-        query = "\"\"".equals(query) ? null : query;
-        int songCount = getIntParameter(request, "songCount", 20);
-        int songOffset = getIntParameter(request, "songOffset", 0);
-        int albumCount = getIntParameter(request, "albumCount", 20);
-        int albumOffset = getIntParameter(request, "albumOffset", 0);
-        int artistCount = getIntParameter(request, "artistCount", 20);
-        int artistOffset = getIntParameter(request, "artistOffset", 0);
-        if (StringUtils.isEmpty(query)) {
-            if (artistCount > 0) {
-                artistService.getArtists(musicFolders, artistCount, artistOffset).forEach(artist -> searchResult.getArtist().add(jaxbContentService.createJaxbArtist(new ArtistID3(), artist, username)));
-            }
-            if (albumCount > 0) {
-                albumService.getAlbums(musicFolders, albumCount, albumOffset).forEach(album -> searchResult.getAlbum().add(jaxbContentService.createJaxbAlbum(new AlbumID3(), album, username)));
-            }
-            if (songCount > 0) {
-                mediaFileService.getSongs(musicFolders, songCount, songOffset).forEach(song -> searchResult.getSong().add(jaxbContentService.createJaxbChild(player, song, username)));
-            }
-        } else {
-            SearchCriteria criteria = new SearchCriteria();
-            criteria.setQuery(StringUtils.trimToEmpty(query));
-            criteria.setCount(artistCount);
-            criteria.setOffset(artistOffset);
-            org.airsonic.player.domain.SearchResult result = searchService.search(criteria, musicFolders, IndexType.ARTIST_ID3);
-            for (org.airsonic.player.domain.Artist artist : result.getArtists()) {
-                searchResult.getArtist().add(jaxbContentService.createJaxbArtist(new ArtistID3(), artist, username));
-            }
-
-            criteria.setCount(albumCount);
-            criteria.setOffset(albumOffset);
-            result = searchService.search(criteria, musicFolders, IndexType.ALBUM_ID3);
-            for (Album album : result.getAlbums()) {
-                searchResult.getAlbum().add(jaxbContentService.createJaxbAlbum(new AlbumID3(), album, username));
-            }
-
-            criteria.setCount(songCount);
-            criteria.setOffset(songOffset);
-            result = searchService.search(criteria, musicFolders, IndexType.SONG);
-            for (MediaFile song : result.getMediaFiles()) {
-                searchResult.getSong().add(jaxbContentService.createJaxbChild(player, song, username));
-            }
-        }
-
-        Response res = createResponse();
-        res.setSearchResult3(searchResult);
-        jaxbWriter.writeResponse(request, response, res);
-    }
 
     @RequestMapping({"/getPlaylists", "/getPlaylists.view"})
     public void getPlaylists(HttpServletRequest request, HttpServletResponse response) {
